requirements/local.txt
-r base.txt

psycopg2-binary
django-debug-toolbar
